axe.spec.js -> headless-chrome.spec.js, factor out axe-tester.js.
